Modified acq2106_tr.py to configure the signals and sources. Now _ACAQ400_TR_BASE is back to its original basics
